@@ -1,6 +1,6 @@
 package com.openclassrooms.realestatemanager.features.detail;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.openclassrooms.realestatemanager.R;
