@@ -2,14 +2,14 @@ package com.openclassrooms.realestatemanager.features.detail
 
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.openclassrooms.realestatemanager.R
 
 
-class DetailFragment : Fragment() {
+class DetailFragment : androidx.fragment.app.Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
